firmware: Prepare for next boot stage on ready-to-boot request
Once the ready-to-boot response is sent, the AP needs to prepare for the
next boot stage. This involves destroying previous connections and
creating them again, which will also fetch the new manifest.

Because the bootloader on the module is updated, there is no point
sending the disconnected events for the earlier control/firmware
protocols and so they should be skipped.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>