connection: destroy svc connection on connection exit or errors
While initializing a connection, the AP requests the SVC to create a
connection between a cport on AP and a cport on the Module.

The opposite of that is missing, when connection is destroyed or if
errors occur after creating the connection. Fix it.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>

---
V3->V4:
- Fix checkpatch warning by breaking function prototype line into two
  lines.

V2->V3:
- Do these operations directly from gb_connection_{init|exit}().

V1->V2:
- This is separate out from a bigger patch, that was part of V1.