manifest: release cport descriptors to avoid 'excess descriptors' warning
If we fail to initialize a cport of a bundle, we abort the entire
bundle. But that leads to following (unnecessary) warnings as few of the
cport descriptors, belonging to the aborted bundle were never parsed:

"greybus: excess descriptors in interface manifest"

Fix that by releasing all cport descriptors for the aborted bundle.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>